자료형 범위 초과로 인한 에러로, Long으로 정의해서 해결

+++ 2021JavaAlgorithm/src/Programmers/Level2/LeastCommon.java
@@ -6,12 +6,12 @@ package Programmers.Level2;
  * arr의 길이 = 15 이하이고, 원소는 100 이하인 자연수.
  * 원소가 100임을 가정할 때 이중 for문으로 풀면 시간초과가 날 수 있을 것 같다.
  *
- * i를 순회할 때, 1부터 하기 때문에 multiple+1 을 해줘야 함. 
+ * i를 순회할 때, 1부터 하기 때문에 multiple+1 을 해줘야 함.
  */
 public class LeastCommon {
   public int solution(int[] arr) {
     int answer = 0;
-    int multiple = 1; //배수를 초기화
+    long multiple = 1; //배수를 초기화
     for(int i=0;i<arr.length;i++) {
       multiple = multiple*arr[i];
     }
